+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/enumerated/EnumeratedSmokeTest.java
@@ -78,7 +78,7 @@ public class EnumeratedSmokeTest extends BaseUnitTestCase {
 		assertThat( hibernateMappingEnumType.isOrdinal(), is(expectedJpaEnumType==EnumType.ORDINAL) );
 		final int expectedJdbcTypeCode = jdbcRegistry.getDescriptor(
 				expectedJpaEnumType == EnumType.ORDINAL ?
-						Types.TINYINT :
+						Types.SMALLINT :
 						Types.VARCHAR
 		).getJdbcTypeCode();
 		assertThat(
